pyproject.toml
[tool.poetry]
name = "homework-project-agata"
version = "0.1.0"
description = ""
authors = ["Agata <agagor1961@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.0.0"
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py
)
'''

[tool.isort]
line_length = 119
